Refactor: move standard-responses into /util/ directory.
